# packages/shared-types/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoTradeIL API",
    "description": "Backend for AutoTradeIL - car trading platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dealers/me": {
      "get": {
        "tags": [
          "dealers"
        ],
        "summary": "Get My Dealer Profile",
        "description": "Return the dealer profile for the currently authenticated user.",
        "operationId": "get_my_dealer_profile_api_v1_dealers_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "dealers"
        ],
        "summary": "Update My Dealer Profile",
        "description": "Free update of dealer-editable fields (Phase 4.4).\n\n`business_name`, `city`, `phone`, `description`, and 3 notification\npreferences are dealer-owned. business_id, license_number, contact_name\nrequire an admin operation and are NOT changeable here.",
        "operationId": "update_my_dealer_profile_api_v1_dealers_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DealerProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DealerProfileUpdate": {
        "properties": {
          "business_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Name"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 15,
                "minLength": 9
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "notification_offers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Offers"
          },
          "notification_deals": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Deals"
          },
          "notification_updates": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Updates"
          }
        },
        "type": "object",
        "title": "DealerProfileUpdate",
        "description": "Free-update fields the dealer can change without admin re-approval."
      },
...
      "DealerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "business_id": {
            "type": "string",
            "title": "Business Id"
          },
          "license_number": {
            "type": "string",
            "title": "License Number"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "lot_size": {
            "type": "integer",
            "title": "Lot Size"
          },
          "contact_name": {
            "type": "string",
            "title": "Contact Name"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "rejected_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected At"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "trust_score": {
            "type": "string",
            "title": "Trust Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "suspended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspended At"
          },
          "suspended_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspended Reason"
          },
          "suspension_silent": {
            "type": "boolean",
            "title": "Suspension Silent",
            "default": false
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "archived_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived Reason"
          },
          "notification_offers": {
            "type": "boolean",
            "title": "Notification Offers",
            "default": true
          },
          "notification_deals": {
            "type": "boolean",
            "title": "Notification Deals",
            "default": true
          },
          "notification_updates": {
            "type": "boolean",
            "title": "Notification Updates",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "business_name",
          "business_id",
          "license_number",
          "phone",
          "city",
          "lot_size",
          "contact_name",
          "verified",
          "verified_at",
          "rejection_reason",
          "rejected_at",
          "tier",
          "trust_score",
          "created_at"
        ],
        "title": "DealerResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}